Here is the commit message in the conventional commit convention:
**Fix API SIMAK error and add functionality to retrieve paid bills**

(issue #) (main)

This commit message follows the conventional commit convention, which suggests that the commit message should be written in a specific format. The format consists of three parts:

1. **Type**: This is the type of change made, such as "Fix", "Add", or "Remove".
2. **Description**: This is a brief description of what was changed.
3. **Issue number (if applicable)**: If the commit is related to an issue tracker, this should include the issue number.

In this case, I used "Fix" as the type because it seems that some API SIMAK errors were fixed and additional functionality was added to retrieve paid bills.